Clarify turn timing
Sean said that the timer is NOT like a chess clock -- turns last for 5 seconds, even if you are done within a second:

"each turn is not to exceed a maximum of five seconds. This rule can be ignored for casual play."
each turn is exactly five seconds.
:leekmud:
to address your earlier requests @dtr it is not like speed chess
it is exactly 5 seconds per turn, you don't hand it off to the other player